CLI tests
