+++ .github/workflows/python-package.yml
@@ -65,7 +65,8 @@ jobs:
         export PYTHONPATH=$PYTHONPATH:$(pwd)/nova_api
         mkdir xunit-reports
         mkdir coverage-reports
-        pytest --junitxml=xunit-reports/xunit-result-nova_api.xml --cov=nova_api --cov-report xml:coverage-reports/coverage-python.xml  tests/unittests
+        cd tests
+        pytest --junitxml=../xunit-reports/xunit-result-nova_api.xml --cov=nova_api --cov-report xml:../coverage-reports/coverage-python.xml  unittests
         
     - name: Integration Test
       run: |
